feat: useFindValidIcon 훅 추가
